+++ keyboards/idobao/id75/id75.c
@@ -13,4 +13,5 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "idobo.h"
+
+#include "id75.h"
