+++ keyboards/idobao/id75/keymaps/greenshadowmaker/rules.mk
