+++ keyboards/idobao/id75/readme.md
@@ -1,0 +1,21 @@
+# IDOBAO ID75
+
+![Idobao ID75](https://cdn.shopify.com/s/files/1/0382/1348/4675/products/02e6d3a5be29c23cb8a93ea203e6da7d.png?v=1621943076)
+
+**75Keys Aluminum Hot-Swappable Ortholinear Mechanical Keyboard.**
+
+Firmware here originally appeared on qmkeyboard.cn
+
+* Keyboard Maintainer: [MechMerlin](https://github.com/mechmerlin)  
+* Hardware Supported: IDOBAO ID75
+* Hardware Availability: [IDOBAO on Drop](https://drop.com/buy/id75-hot-swappable-ortholinear-keyboard-kit) or [IDOBAO direct](https://www.idobao.net/collections/keyboard-kit/products/idobao-id75v1-hot-swappable-mechanical-keyboard-kit)
+
+Make example for this keyboard (after setting up your build environment):
+
+    make idobao/id75:default
+
+or
+
+    qmk compile -kb idobao/id75 -km default
+        
+See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
