--- keyboards/idobo/readme.md
@@ -1,17 +1,0 @@
-# IDOBO
-
-![IDOBO](imgur.com image replace me!)
-
-75Keys Aluminum Mechanical Keyboard.
-
-Firmware here originally appeared on qmkeyboard.cn
-
-Keyboard Maintainer: [MechMerlin](https://github.com/mechmerlin)  
-Hardware Supported: IDOBO PCB
-Hardware Availability: [IDOBO on Drop](https://drop.com/buy/id75-hot-swappable-ortholinear-keyboard-kit)
-
-Make example for this keyboard (after setting up your build environment):
-
-    make idobo:default
-
-See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
